refactor curves
