+++ src/dgcv/curves/anonymizer.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import numpy as np
 
 from dgcv.configuration.cfg import config
-from dgcv.curves.importer import CurvesImporter
+from dgcv.curves.importer.importer import CurvesImporter
 from dgcv.sigpro.sigpro import lowpass_filter
 
 NOISE_DAMPING = 100
