+++ src/dgcv/curves/importer/reader.py
